Further refactorization
